# lobel_store_frontend/docs/openapi.generated.yml
swagger: '2.0'
info:
  title: Lobel Store API
  description: Documentation API pour Lobel Store
  version: v1
basePath: /api
consumes:
- application/json
produces:
- application/json
securityDefinitions:
  Basic:
    type: basic
security:
- Basic: []
paths:
...
  /orders/orders/cart/:
    get:
      operationId: orders_orders_cart
      description: Retourne le panier actif (commande non finalisée) du client connecté.
      parameters:
      - name: page
        in: query
        description: A page number within the paginated result set.
        required: false
        type: integer
      - name: page_size
        in: query
        description: Number of results to return per page.
        required: false
        type: integer
      responses:
        '200':
          description: ''
          schema:
            required:
            - count
            - results
            type: object
            properties:
              count:
                type: integer
              next:
                type: string
                format: uri
                x-nullable: true
              previous:
                type: string
                format: uri
                x-nullable: true
              results:
                type: array
                items:
                  $ref: '#/definitions/OrderRead'
      tags:
      - orders
    parameters: []
...
  /orders/orders/cart/merge/:
    post:
      operationId: orders_orders_cart_merge_cart
      description: ''
      parameters:
      - name: data
        in: body
        required: true
        schema:
          $ref: '#/definitions/OrderRead'
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/OrderRead'
      tags:
      - orders
    parameters: []
  /orders/orders/checkout/create-order/:
    post:
      operationId: orders_orders_checkout_checkout_create_order
      description: ''
      parameters:
      - name: data
        in: body
        required: true
        schema:
          $ref: '#/definitions/OrderRead'
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/OrderRead'
      tags:
      - orders
    parameters: []
...
  /orders/orders/checkout/pending/:
    get:
      operationId: orders_orders_checkout_checkout_pending
      description: ''
      parameters:
      - name: page
        in: query
        description: A page number within the paginated result set.
        required: false
        type: integer
      - name: page_size
        in: query
        description: Number of results to return per page.
        required: false
        type: integer
      responses:
        '200':
          description: ''
          schema:
            required:
            - count
            - results
            type: object
            properties:
              count:
                type: integer
              next:
                type: string
                format: uri
                x-nullable: true
              previous:
                type: string
                format: uri
                x-nullable: true
              results:
                type: array
                items:
                  $ref: '#/definitions/OrderRead'
      tags:
      - orders
    parameters: []
  /orders/orders/checkout/preview/:
    post:
      operationId: orders_orders_checkout_checkout_preview
      description: ''
      parameters:
      - name: data
        in: body
        required: true
        schema:
          $ref: '#/definitions/OrderRead'
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/OrderRead'
      tags:
      - orders
    parameters: []
...
  /orders/orders/{id}/receipt/:
    get:
      operationId: orders_orders_receipt
      description: ''
      parameters: []
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/OrderRead'
      tags:
      - orders
    parameters:
    - name: id
      in: path
      description: A unique integer value identifying this order.
      required: true
      type: integer
  /orders/orders/{id}/request-refund/:
    post:
      operationId: orders_orders_request_refund
      description: ''
      parameters:
      - name: data
        in: body
        required: true
        schema:
          $ref: '#/definitions/OrderRead'
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/OrderRead'
      tags:
      - orders
    parameters:
    - name: id
      in: path
      description: A unique integer value identifying this order.
      required: true
      type: integer
...
definitions:
...
  OrderItem:
    required:
    - variant_id
    - quantity
    type: object
    properties:
      id:
        title: ID
        type: integer
        readOnly: true
      product_id:
        title: Product id
        type: integer
        readOnly: true
      product_reference:
        title: Product reference
        type: integer
        readOnly: true
        x-nullable: true
      product_name:
        title: Product name
        type: string
        readOnly: true
        minLength: 1
      variant_id:
        title: Variant id
        type: integer
      variant_reference:
        title: Variant reference
        type: integer
        readOnly: true
        x-nullable: true
      variant_name:
        title: Variant name
        type: string
        readOnly: true
        minLength: 1
      color:
        title: Color
        type: string
        readOnly: true
        minLength: 1
      size:
        title: Size
        type: string
        readOnly: true
        minLength: 1
      sku:
        title: Sku
        type: string
        readOnly: true
        minLength: 1
      quantity:
        title: Quantity
        type: integer
        maximum: 99
        minimum: 1
      unit_price:
        title: Unit price
        type: string
        format: decimal
        readOnly: true
      currency:
        title: Currency
        type: string
        readOnly: true
        minLength: 1
        x-nullable: true
      discount_amount:
        title: Discount amount
        type: string
        format: decimal
        readOnly: true
      subtotal:
        title: Subtotal
        type: string
        format: decimal
        readOnly: true
        x-nullable: true
      line_total:
        title: Line total
        type: string
        format: decimal
        readOnly: true
      date_added:
        title: Date added
        type: string
        format: date-time
        readOnly: true
...
  User:
    required:
    - username
    type: object
    properties:
      id:
        title: ID
        type: integer
        readOnly: true
      username:
        title: Username
        description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
          only.
        type: string
        pattern: ^[\w.@+-]+$
        maxLength: 150
        minLength: 1
      first_name:
        title: First name
        type: string
        maxLength: 150
      last_name:
        title: Last name
        type: string
        maxLength: 150
      email:
        title: Email address
        type: string
        format: email
        maxLength: 254
      is_active:
        title: Active
        description: Designates whether this user should be treated as active. Unselect
          this instead of deleting accounts.
        type: boolean
  CustomerRead:
    type: object
    properties:
      id:
        title: ID
        type: integer
        readOnly: true
      user:
        $ref: '#/definitions/User'
      country:
        title: Country
        description: 'Code pays ISO (ex: FR, ML)'
        type: string
        readOnly: true
        minLength: 1
      phone_number:
        title: Phone number
        description: Numéro de téléphone au format international
        type: string
        readOnly: true
        minLength: 1
      address:
        title: Address
        type: string
        readOnly: true
        minLength: 1
      date_created:
        title: Date created
        type: string
        format: date-time
        readOnly: true
  OrderRead:
    type: object
    properties:
      id:
        title: ID
        type: integer
        readOnly: true
      customer:
        $ref: '#/definitions/CustomerRead'
      date_ordered:
        title: Date ordered
        type: string
        format: date-time
        readOnly: true
      complete:
        title: Complete
        type: boolean
        readOnly: true
      status:
        title: Status
        type: string
        enum:
        - cart
        - pending_payment
        - payment_processing
        - payment_failed
        - paid
        - preparing
        - shipped
        - delivered
        - cancelled
        - expired
        - refund_required
        - refund_pending
        - refunded
        - refund_failed
        readOnly: true
      paid_at:
        title: Paid at
        type: string
        format: date-time
        readOnly: true
        x-nullable: true
      transaction_id:
        title: Transaction id
        type: string
        readOnly: true
        minLength: 1
        x-nullable: true
      items:
        type: array
        items:
          $ref: '#/definitions/OrderItem'
        readOnly: true
      cart_total:
        title: Cart total
        type: string
        format: decimal
        readOnly: true
      cart_items:
        title: Cart items
        type: integer
        readOnly: true
      snapshot_at:
        title: Snapshot at
        type: string
        format: date-time
        readOnly: true
        x-nullable: true
      customer_name:
        title: Customer name
        type: string
        readOnly: true
        minLength: 1
      customer_email:
        title: Customer email
        type: string
        format: email
        readOnly: true
        minLength: 1
      delivery_recipient_name:
        title: Delivery recipient name
        type: string
        readOnly: true
        minLength: 1
      delivery_phone:
        title: Delivery phone
        type: string
        readOnly: true
        minLength: 1
      delivery_address:
        title: Delivery address
        type: string
        readOnly: true
        minLength: 1
      delivery_country:
        title: Delivery country
        type: string
        readOnly: true
        minLength: 1
      delivery_region:
        title: Delivery region
        type: string
        readOnly: true
        minLength: 1
      delivery_city:
        title: Delivery city
        type: string
        readOnly: true
        minLength: 1
      delivery_district:
        title: Delivery district
        type: string
        readOnly: true
        minLength: 1
      delivery_street:
        title: Delivery street
        type: string
        readOnly: true
        minLength: 1
      delivery_instructions:
        title: Delivery instructions
        type: string
        readOnly: true
        minLength: 1
      delivery_method_code:
        title: Delivery method code
        type: string
        readOnly: true
        minLength: 1
      delivery_method_label:
        title: Delivery method label
        type: string
        readOnly: true
        minLength: 1
      delivery_eta_min_days:
        title: Delivery eta min days
        type: integer
        readOnly: true
        x-nullable: true
      delivery_eta_max_days:
        title: Delivery eta max days
        type: integer
        readOnly: true
        x-nullable: true
      billing_same_as_shipping:
        title: Billing same as shipping
        type: boolean
        readOnly: true
      billing_address:
        title: Billing address
        type: string
        readOnly: true
        minLength: 1
      checkout_version:
        title: Checkout version
        type: string
        readOnly: true
        minLength: 1
      subtotal_amount:
        title: Subtotal amount
        type: string
        format: decimal
        readOnly: true
        x-nullable: true
      shipping_amount:
        title: Shipping amount
        type: string
        format: decimal
        readOnly: true
      discount_amount:
        title: Discount amount
        type: string
        format: decimal
        readOnly: true
      total_amount:
        title: Total amount
        type: string
        format: decimal
        readOnly: true
        x-nullable: true
      currency:
        title: Currency
        type: string
        readOnly: true
        minLength: 1
      preparation_started_at:
        title: Preparation started at
        type: string
        format: date-time
        readOnly: true
        x-nullable: true
      shipped_at:
        title: Shipped at
        type: string
        format: date-time
        readOnly: true
        x-nullable: true
      delivered_at:
        title: Delivered at
        type: string
        format: date-time
        readOnly: true
        x-nullable: true
      cancelled_at:
        title: Cancelled at
        type: string
        format: date-time
        readOnly: true
        x-nullable: true
      refund_requested_at:
        title: Refund requested at
        type: string
        format: date-time
        readOnly: true
        x-nullable: true
      refunded_at:
        title: Refunded at
        type: string
        format: date-time
        readOnly: true
        x-nullable: true
      stock_consumed_at:
        title: Stock consumed at
        type: string
        format: date-time
        readOnly: true
        x-nullable: true
      stock_released_at:
        title: Stock released at
        type: string
        format: date-time
        readOnly: true
        x-nullable: true
      status_history:
        title: Status history
        type: string
        readOnly: true
      payment_processing_at:
        title: Payment processing at
        type: string
        format: date-time
        readOnly: true
        x-nullable: true
      payment_failed_at:
        title: Payment failed at
        type: string
        format: date-time
        readOnly: true
        x-nullable: true
      expired_at:
        title: Expired at
        type: string
        format: date-time
        readOnly: true
        x-nullable: true
      status_label:
        title: Status label
        type: string
        readOnly: true
        minLength: 1
      timeline:
        title: Timeline
        type: string
        readOnly: true
      available_actions:
        title: Available actions
        type: string
        readOnly: true
      payment:
        title: Payment
        type: string
        readOnly: true
...